+++ src/ShapeCrawler/Drawing/ShapeFillImage.cs
@@ -51,21 +51,5 @@ internal sealed class ShapeFillImage : IImage
         this.Update(sourceBytes);
     }
 
-    public byte[] AsByteArray()
-    {
-        var stream = this.sdkImagePart.GetStream();
-        var mStream = new MemoryStream();
-        var buffer = new byte[1024];
-
-        int read;
-
-        while ((read = stream.Read(buffer, 0, buffer.Length)) > 0)
-        {
-            mStream.Write(buffer, 0, read);
-        }
-
-        stream.Close();
-
-        return mStream.ToArray();
-    }
+    public byte[] AsByteArray() => new WrappedImagePart(this.sdkImagePart).AsBytes(); 
 }